requirements.txt
# Core and GUI dependencies
setuptools>=65.0.0
wheel>=0.38.0
pillow>=10.0.0

# Monitor Control
monitorcontrol>=4.1.0

# Samsung Local Control (Tizen)
samsungtvws>=2.6.0
requests>=2.31.0
websocket-client>=1.6.0

# Build
pyinstaller>=6.0.0
